feat: migrate to otterbrix b1 + LIMIT pushdown + Spark Connect E2E
otterbrix a13 -> 1.0.0b1 source migration (deps were bumped in c7deeda):
- flat logical-plan nodes: node_drop_t/drop_target_kind and
  node_catalog_resolve_t/resolve_kind replace the split node types
  (parser, name_resolution, sql generator, catalog_manager)
- multi-chunk cursor: session_payload now carries
  std::pmr::vector<data_chunk_t> (<=1024 rows/chunk) instead of one
  uncapped chunk; all frontends (mysql/pg/flight/spark) and tests iterate
  chunks; fixes a latent >1024-row truncation (new repro test)

Fixes uncovered along the way:
- LIMIT/OFFSET are now pushed down to remote backends (generate_select
  dropped the node_limit_t child, so LIMIT n returned all rows)
- Spark Connect: read query from SqlCommand.input (Spark 4.0), kToDf oneof
  rename, transform_result insert_rows_t ctor, result_encoder alias()
  segfault on unaliased leaf types, grpcpp/support/status.h include

Tooling:
- Dockerfile.test: cppstd=17 pins for protobuf/re2 (abseil cluster)
- Dockerfile.spark-test: pyspark[connect] + pandas
- docker-run-tests.sh: build+run test-spark-client

Rule cleanup in touched files: remove unused utility/shared_flight_data.hpp,
dynamic_cast->static_cast, thread_local RNG rehome, 64-bit PG row count,
drop <tuple>/std::ignore/(void) casts.

+++ frontend/postgres_server/packet/packet_utils.hpp
@@ -33,11 +33,11 @@ namespace frontend::postgres {
     };
 
     struct command_complete_tag {
-        static command_complete_tag simple_command(NodeTag node, int32_t rows = 0);
-        static command_complete_tag select(int32_t rows = 0);
-        static command_complete_tag insert(int32_t rows = 0);
-        static command_complete_tag update(int32_t rows = 0);
-        static command_complete_tag delete_rows(int32_t rows = 0);
+        static command_complete_tag simple_command(NodeTag node, int64_t rows = 0);
+        static command_complete_tag select(int64_t rows = 0);
+        static command_complete_tag insert(int64_t rows = 0);
+        static command_complete_tag update(int64_t rows = 0);
+        static command_complete_tag delete_rows(int64_t rows = 0);
         static command_complete_tag begin();
         static command_complete_tag commit();
         static command_complete_tag rollback();
